Create RF-Chain-simuation
This is the python code for RF Chain simulation of GRAND (Xidian version). The operation of it need the data of the antenna and other S parameters of devices. But the size of the data beyond the limitation of Github, hence It is wellcome to send me email (zhangpf@mail.xidian.edu.cn) and I can send you the reated data used in this code.